+++ src/graphic/buffer/IndexBuffer.cpp
@@ -9,11 +9,11 @@ IndexBuffer::IndexBuffer()
 {
 }
 
-void IndexBuffer::Initialize(ID3D11Device* device, const uint32* indices, uint32 numIndices)
+void IndexBuffer::Initialize(ID3D11Device* device, const uint32* indices, uint64 numIndices)
 {
     D3D11_BUFFER_DESC ibd;
     ibd.Usage = D3D11_USAGE_IMMUTABLE;
-    ibd.ByteWidth = sizeof(uint32) * numIndices;
+    ibd.ByteWidth = static_cast<uint32>(sizeof(uint32) * numIndices);
     ibd.BindFlags = D3D11_BIND_INDEX_BUFFER;
     ibd.CPUAccessFlags = 0;
     ibd.MiscFlags = 0;
@@ -26,7 +26,7 @@ void IndexBuffer::Initialize(ID3D11Device* device, const uint32* indices, uint32
 
 void IndexBuffer::Initialize(ID3D11Device* device, const std::vector<uint32> indices)
 {
-    Initialize(device, &indices[0], indices.size());
+    Initialize(device, &indices[0], static_cast<uint32>(indices.size()));
 }
 
 OC_NS_END;
